tg client great again

+++ smarthouse/scenarios/light_scenarios.py
@@ -67,11 +67,13 @@ async def tg_actions():
     await tg_client.update_tg()
 
 
-@looper(0.1)
+@looper(3)
 async def clear_tg():
     tg_client = TGClient()
 
     to_delete_timestamp, message_id = await tg_client.to_delete_messages.get()
+    if time.time() - to_delete_timestamp < 5:
+        return 0.1
     if to_delete_timestamp <= time.time():
         await tg_client.delete_message(int(message_id))
     else:
